requirements-ml.txt
# Dependências do pipeline de ML — Delfos v1.0
# Instalar com: pip install -r requirements-ml.txt

# Core ML
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0

# Tracking de experimentos
mlflow>=2.12.0

# Dados
pandas>=2.2.0
pyarrow>=15.0.0     # suporte a parquet
numpy>=1.26.0

# StatsBomb open data
statsbombpy>=1.0.3

# Visualização
matplotlib>=3.8.0
seaborn>=0.13.0

# Serialização / ONNX (US-4.3)
joblib>=1.3.0
onnx>=1.16.0
skl2onnx>=1.16.0
onnxmltools>=1.12.0
onnxruntime>=1.17.0

# API de inferência (US-4.4)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
scipy>=1.12.0
python-dotenv>=1.0.0

# Testes da API
httpx>=0.27.0
pytest-asyncio>=0.23.0
